Tensorflow 1.x/requirements.txt
tensorflow-gpu==1.15
keras==2.2.4
tf-slim==1.1.0
scipy==1.2.1
pillow
